My idea for future backwards compatibility support
now its time to write the code :). original plan was to shove an entire type model in the header but that design is pretty flawed. Its impossible to warmup and might lead to huge runtime hiccups. could lead to security exploits too. I think the key + size is a better solution